fix(expense): fix problem with amount column by switching from integer to float

+++ app/models/expense_model.py
@@ -1,6 +1,6 @@
 import enum
 from datetime import datetime, timezone
-from sqlalchemy import Integer, String, Column, Enum, DateTime, ForeignKey
+from sqlalchemy import Integer, Float, String, Column, Enum, DateTime, ForeignKey
 from app.db.database import Base
 
 class ExpenseCategory(str, enum.Enum):
@@ -16,7 +16,7 @@ class Expense(Base):
 
   id = Column(Integer, primary_key=True)
   description = Column(String, nullable=False, index=True)
-  amount = Column(Integer, nullable=False, index=True)
+  amount = Column(Float, nullable=False, index=True)
   date = Column(DateTime, default=lambda: datetime.now(timezone.utc), index=True)
   category = Column(Enum(ExpenseCategory), index=True, nullable=False)
   updated_at = Column(DateTime, default=lambda: datetime.now(timezone.utc), onupdate=lambda: datetime.now(timezone.utc))
